First version of route layout
