+++ app/layout.js
@@ -1,0 +1,24 @@
+import './globals.css'
+import MainFooter from '../components/MainFooter'
+
+export default function RootLayout({ children }) {
+    return (
+        <html lang="en">
+            <head>
+                <title>Create Next App</title>
+                <meta
+                    name="description"
+                    content="Generated by create next app"
+                />
+                <link rel="icon" href="/favicon.ico" />
+            </head>
+            <body>
+                <header></header>
+                <main>{children}</main>
+                <footer>
+                    <MainFooter />
+                </footer>
+            </body>
+        </html>
+    )
+}
