+++ app/page.js
@@ -1,0 +1,12 @@
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../styles/Home.module.css'
+
+export default function Home() {
+    return (
+        <div className={styles.container}>
+            <aside></aside>
+            <h1>Main</h1>
+        </div>
+    )
+}
