+++ app/princesses/page.js
@@ -1,0 +1,9 @@
+export default function Princesses() {
+    return (
+        <div>
+            <main>
+                <main>Princesses</main>
+            </main>
+        </div>
+    )
+}
