+++ components/MainFooter.js
@@ -1,0 +1,7 @@
+import classes from './MainFooter.module.css'
+
+function MainFooter() {
+    return <section className={classes.footer}>Footer</section>
+}
+
+export default MainFooter
