+++ next.config.js
@@ -1,5 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+    experimental: { appDir: true },
     reactStrictMode: true,
     swcMinify: true
 }
